Incorporated file path bundling back into load_api_key function
